Rename wrist commands
