+++ src/main/java/org/team1540/robot2019/commands/groups/GetHatchFloor.java
@@ -4,7 +4,7 @@ import edu.wpi.first.wpilibj.command.Command;
 import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.team1540.robot2019.Robot;
 import org.team1540.robot2019.commands.elevator.MoveElevatorToZero;
-import org.team1540.robot2019.commands.wrist.LowerWrist;
+import org.team1540.robot2019.commands.wrist.WristDown;
 import org.team1540.rooster.util.SimpleCommand;
 
 public class GetHatchFloor extends CommandGroup {
@@ -13,7 +13,7 @@ public class GetHatchFloor extends CommandGroup {
     addSequential(new SimpleCommand("in", Robot.hatch::slideIn, Robot.hatch));
     addSequential(new SimpleCommand("close", Robot.hatch::release, Robot.hatch));
     addSequential(new MoveElevatorToZero());
-    addSequential(new LowerWrist());
+      addSequential(new WristDown());
     addSequential(new SimpleCommand("out", Robot.hatch::slideOut, Robot.hatch));
     addSequential(new Command() {
       @Override
