+++ src/main/java/org/team1540/robot2019/commands/groups/IntakeSequence.java
@@ -4,7 +4,7 @@ import edu.wpi.first.wpilibj.command.CommandGroup;
 import org.team1540.robot2019.Robot;
 import org.team1540.robot2019.commands.cargo.AutoIntakeCargo;
 import org.team1540.robot2019.commands.elevator.MoveElevatorToZero;
-import org.team1540.robot2019.commands.wrist.LowerWrist;
+import org.team1540.robot2019.commands.wrist.WristDown;
 import org.team1540.rooster.util.SimpleCommand;
 
 public class IntakeSequence extends CommandGroup {
@@ -13,6 +13,6 @@ public class IntakeSequence extends CommandGroup {
         addParallel(new SimpleCommand("alaska's", Robot.hatch::slideIn, Robot.hatch));
         addSequential(new MoveElevatorToZero());
         addParallel(new AutoIntakeCargo());
-        addSequential(new LowerWrist());
+        addSequential(new WristDown());
     }
 }
