+++ src/main/java/org/team1540/robot2019/commands/wrist/WristUp.java
@@ -1,20 +1,24 @@
 package org.team1540.robot2019.commands.wrist;
 
-import edu.wpi.first.wpilibj.command.ConditionalCommand;
+import edu.wpi.first.wpilibj.command.Command;
 import org.team1540.robot2019.Robot;
 import org.team1540.robot2019.Tuning;
-import org.team1540.rooster.util.SimpleLoopCommand;
 
-public class WristUp extends ConditionalCommand {
+public class WristUp extends Command {
 
-    public WristUp() {
-        super(new MoveWristUp(),
-            new SimpleLoopCommand("Hold Wrist", () -> Robot.wrist.set(Tuning.wristHoldThrot),
-                Robot.wrist));
+    @Override
+    protected void initialize() {
+        Robot.wrist.clearMidFlag();
+        Robot.wrist.set(Tuning.wristUpTravelThrot);
     }
 
     @Override
-    protected boolean condition() {
-        return Robot.wrist.isAtBtm();
+    protected void end() {
+        Robot.wrist.set(Tuning.wristHoldThrot);
+    }
+
+    @Override
+    protected boolean isFinished() {
+        return Robot.wrist.getMidFlag();
     }
 }
